#240 Refactor: auth 관련 서비스 분리 및 회원가입 method 수정
